requirements.txt
python-dotenv
openai
python-telegram-bot==20.6
